Make selections show nicely in inactive windows
The build window doesn't have focus but uses selection to mark the next
error line. As it is inactive, the selection color gets darker,
resulting in almost the same color as the text. This uses a color not
inclued in the solarized palette, but it works well in active and
inactive windows.